Backend/requirements.txt
flask
Flask-SQLAlchemy
jinja2
psycopg
flask-login
python-dotenv
flask-bcrypt
flask-testing